Beta version 0.0.1
package (unpublished)

Update README.md

Add manifest
